feat(borrow): clean up confirm header
