+++ packages/blockchain-wallet-v4-frontend/src/modals/Borrow/BorrowForm/template.success.tsx
@@ -75,17 +75,6 @@ const FiatContainer = styled.div`
   background-color: ${props => props.theme.grey000};
 `
 
-const ErrorText = styled(Text)`
-  display: inline-flex;
-  font-weight: 500;
-  font-size: 14px;
-  padding: 6px 12px;
-  border-radius: 32px;
-  background-color: ${props => props.theme.red000};
-  color: ${props => props.theme.red800};
-  margin-bottom: 16px;
-`
-
 const ButtonContainer = styled.div`
   display: flex;
   justify-content: space-between;
@@ -204,16 +193,6 @@ const Success: React.FC<InjectedFormProps<{}, Props> & Props> = props => {
             collateral={0}
             displayName={displayName}
           />
-          {props.error && (
-            <ErrorText>
-              <Icon
-                name='alert-filled'
-                color='red600'
-                style={{ marginRight: '4px' }}
-              />
-              Error: {props.error}
-            </ErrorText>
-          )}
           <ButtonContainer>
             <Button
               nature='empty'
@@ -231,18 +210,14 @@ const Success: React.FC<InjectedFormProps<{}, Props> & Props> = props => {
               nature='primary'
               type='submit'
               data-e2e='borrowSubmit'
-              disabled={props.submitting || props.invalid}
+              disabled={props.invalid}
             >
-              {props.submitting ? (
-                <HeartbeatLoader height='16px' width='16px' color='white' />
-              ) : (
-                <Text size='16px' weight={600} color='white'>
-                  <FormattedMessage
-                    id='modals.borrow.collateralform.create'
-                    defaultMessage='Create Loan'
-                  />
-                </Text>
-              )}
+              <Text size='16px' weight={600} color='white'>
+                <FormattedMessage
+                  id='modals.borrow.collateralform.continue'
+                  defaultMessage='Continue'
+                />
+              </Text>
             </Button>
           </ButtonContainer>
         </>
