+++ packages/blockchain-wallet-v4-frontend/src/modals/Borrow/ConfirmBorrow/index.tsx
@@ -29,6 +29,10 @@ type Props = OwnProps & LinkDispatchPropsType & LinkStatePropsType
 class ConfirmBorrow extends PureComponent<Props> {
   state = {}
 
+  componentWillUnmount () {
+    this.props.borrowActions.destroy()
+  }
+
   handleSubmit = () => {
     this.props.borrowActions.createBorrow()
   }
